// apps/web/lib/api/schemas/publicProfile.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Pi-Cast API
 * Pi-Cast API documentation with Better Auth integration
 * OpenAPI spec version: 1.0.0
 */
import type { UserPublic } from './userPublic';
import type { LessonItem } from './lessonItem';

export interface PublicProfile {
  user: UserPublic;
  lessons: LessonItem[];
  isFollowing: boolean;
  isFollowedByCurrentUser: boolean;
}