// apps/web/lib/api/schemas/lessonItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Pi-Cast API
 * Pi-Cast API documentation with Better Auth integration
 * OpenAPI spec version: 1.0.0
 */

export interface LessonItem {
  id: string;
  title: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  thumbnailUrl: string | null;
  /** @nullable */
  level: string | null;
  viewCount: number;
  likeCount: number;
  /** @nullable */
  duration: number | null;
  createdAt: number;
}